Generate autoloads for sub-directories at package build time.
Create an autoloads file that collects autoloads from Elisp files in
sub-directories of ProofGeneral's  top-level directory.  The reason
for generating this file is that the standard Emacs package
installation procedure will only inspect the Elisp files in
ProofGeneral's top-level directory, but will not inspect the
sub-directories (c.f `package-generate-autoloads' in the standard
Emacs distribution).

The `pkg' target of the Makefile now creates a small "shim" to cause
the autoloads collected from sub-directories (described in the
preceeding paragraph) to be loaded whenever ProofGeneral-autoloads.el
file (which is generated by `package-generate-autoloads') is loaded.

Finally, this change adds a couple of autoload "cookies" which were
needed (at least in my testing) so that Coq mode can be used
immediately upon package installation (i.e., in the same Emacs session
in which `package-install <a_location_of_ProofGeneral_package>' was
run.  Without those cookies, I was getting an error complaining that
the function `coq-mode-config' was undefined.  Similar "cookies" may
be needed for other proof assistants (I only use Coq, so I am choosing
to stay conservative in my changes).