prooftree: Update proof tree display support for Coq 8.10
This is a major change, because the new Show Goal command in Coq
8.10 can print goals for previous states and therefore permits to
remove all the code that was tracking goals and existential from
PG and to rely on prooftree to request goals and goal updates.

This change does break compatibility for the proof tree display
for Coq 8.6 and before. (The old Coq versions themselves work,
it's just that proof tree display won't work.)

In paricular:
- delete urgent proof tree actions, the hooks and all the
  corresponding code
- delete sequent hash and existentials info and all the corresponding code
- adopt prooftree regular expressions and show goal functions to
  Coq 8.10 (implemented in coq/coq#10489)
- new hooks proof-tree-start-display-hook and
  proof-tree-stop-display-hook for switching Coq's dependent evar
  line on and off
- forward the new show goal commands from prooftree to Coq
- proof tree display starts with the Lemma command, because Proof
  does not display anything anymore
- deleted superfluous proof tree code in hol-light and
  deconfigured prooftree for Hol-Light, IIRC, it was never
  working. But I am happy to provide support, if anybody wants to
  pick this up.