Preventive support of "goal instead of "subgoal" in coq messages.
Coq team is currently considering changing all messages containing
"subgoal[s]" into "goal[s]". This commit allows for both.

I make this change as early as possible, even if this chage may not
land in coq's master, so that we support this as early as possible.

I don't see how this could break anything but again the earliest the
best.