fix another make magic problem from 2018
In 7389d43893569ff0e1eff892254901876fc8225e
proof-ready-for-assistant was moved to proof-script.el, breaking
make magic in a complicated chain. The auto load for
proof-ready-for-assistant caused proof-script.el to be loaded
before proof-assistant was set, causing
proof-eval-when-ready-for-assistant, to put stuff into
proof-ready-for-assistant-hook. proof-ready-for-assistant loads
pg-custom.el by require, therefore, for the first proof assistant
in the loop in docstring-magic everything was defined when
executing the hooks at the end of proof-ready-for-assistant. The
second assistant caused errors than.

Fixes #542