add test for default-directory
This test exhibits a bug in background compilation that became
apparent when working on tramp support: Strange things may happen
when the user changes the current buffer, because part of the
code may then run with an incorrect default-directory.

Introduce hooks before and after busy waiting in the cct library.