enable compilation for remote files via tramp
tramp support is not complete with this patch, but works if used
with care. There are a few assertions that quick user
interactions can trigger. Automatic background compilation via
tramp is very slow, because tramp opens a new connection with all
overhead for each asynchronous coqdep and coqc invocation. With
this patch, tramp support only works for Emacs version 26 or
later, because it uses make-nearby-temp-file.

Modifications:
- switch to functions that properly deal with remote or magic files
- protect sentinels against multiple instances running
  interleaved, which suddenly happens under tramp
- recognize the error when coqdep cannot access a file

Fixes #104 and fixes #203.