Handle single-quotes in arguments to `-arg`
A solution sketch to issue #589.

coq_makefile parses _CoqProject files in a slightly weird way. If we
write `-arg "a b"` in _CoqProject then it passes the arguments `a` and
`b` to coqc, but if we write `-arg "'a b'"` it passes `a b` to coqc, as
a single argument.
This commit tries to adapt to this behavior.