CI: do not artificially restrict parallel test execution
For whatever reason, the test job always runs late in a CI run (I am
blaming the max-parallel annotation for this). This leads to longer CI
runs, because, at the end, for several rounds on 6 instances of the
test job are executing. Note that Github starts at most 20 of our
tests in parallel, so it is not our responsibility to ensure that
limit.