omit-proofs: split command processing into two phases
The first phase only classifies proofs to be omitted and stuff not to
be omitted. The second phase does the actual replacement and adjusts
the overlays/spans. Feature wise this commit does not change anything
but it enables reusing the first phase for new features.

Additionally, slightly extend the omit proofs test.