proof-check: new feature for listing passing and failing tests
M-x proof-script-check-proofs checks all proofs in one buffer and
displays a list with which tests currently pass or fail. This is a
Proof General implementation of coq/coq#1147. It enables a PVS-like
workflow where files are processed with -vos and
proof-omit-proofs-option enabled such that one can focus on the most
interesting/difficult proof instead of the first failing one.