Snapshot match limits in proof-format function
The existing code had a subtle bug because entries in the alist were
usually lists like:

  '("%l" proof-assistant-format-lambda curvalue)

This looks a little odd at first, but it's the same thing as:

  '("%l" . (proof-assistant-format-lambda curvalue))

We then use eval as a sort of macro evaluation system to figure out
what an occurrence of "%l" should expand to.

The function that gets evaluated might call string-match itself, so
it's important to make sure to extract the match limits *before*
evaluating the function that calculates a replacement.