Match "." and "..." more strictly in coq-indent.el
This was prompted by issue #853, which had the following example that
didn't work properly:

    Axiom Ninterval: nat -> nat -> Prop.
    Notation "[ n  '....'  m ]" := (Ninterval n m).
    Notation "x 'BETWEEN' n 'AND' m " := ([n .... m] x) (at level 1).
    Axiom something_else : Prop.

The problem was that the regex in coq-period-end-command would match
"....", even though it's neither one nor three periods.

Because I found the regexes quite hard to read, this commit also
splits up the concats a bit, which I *think* makes them a bit easier.