Really disabling a feature marked as such.
coq-show (C-c C-a C-s) used to be smart and tried to look at the
cursor position. If the cursos was on the locked region it tried to
extract the goal state as stored at this point. But states are not
stored anymore (coq runs silently now). This was marked as disabled in
comments but the code was not unplugged.

Even if this becomes possible again I think this feature (looking at
the goal at point in file instead of at the end of locked region)
needs to be a different command.